workflowmasterearnix_bot - Copy/requirements.txt
pyrogram
tgcrypto
requests
beautifulsoup4
openpyxl
python-dotenv

